refactor: 모듈 분리 및 JUnit5 적용
