+++ app/build.gradle.kts
@@ -37,6 +37,7 @@ android {
 }
 
 dependencies {
+    implementation(project(":domain"))
     implementation("androidx.core:core-ktx:1.9.0")
     implementation("androidx.appcompat:appcompat:1.6.0")
     implementation("com.google.android.material:material:1.7.0")
