--- app/src/main/java/woowacourse/movie/domain/discount/discountcondition/DiscountCondition.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.domain.discount.discountcondition
-
-import java.time.LocalDateTime
-
-interface DiscountCondition {
-
-    fun isDiscount(dateTime: LocalDateTime): Boolean
-}
