+++ domain/build.gradle.kts
@@ -1,0 +1,28 @@
+
+plugins {
+    kotlin("jvm")
+}
+
+group = "camp.nextstep.edu"
+version = "1.0-SNAPSHOT"
+
+dependencies {
+    testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
+    testImplementation("org.assertj", "assertj-core", "3.22.0")
+    testImplementation("io.kotest", "kotest-runner-junit5", "5.2.3")
+}
+
+tasks {
+    compileKotlin {
+        kotlinOptions.jvmTarget = "11"
+    }
+    compileTestKotlin {
+        kotlinOptions.jvmTarget = "11"
+    }
+    test {
+        useJUnitPlatform()
+    }
+    ktlint {
+        verbose.set(true)
+    }
+}
