+++ domain/src/main/java/woowacourse/movie/discount/discountcondition/DiscountCondition.kt
@@ -1,0 +1,8 @@
+package woowacourse.movie.discount.discountcondition
+
+import java.time.LocalDateTime
+
+interface DiscountCondition {
+
+    fun isDiscount(dateTime: LocalDateTime): Boolean
+}
