+++ domain/src/main/java/woowacourse/movie/discount/discountpolicy/DiscountPolicyAdapter.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.discount.discountpolicy
+
+import woowacourse.movie.model.Money
+import java.time.LocalDateTime
+
+class DiscountPolicyAdapter(private val policies: List<DiscountPolicy>) {
+    fun discount(price: Money, dateTime: LocalDateTime): Money {
+        var discountedMoney = price
+        policies.forEach { discountedMoney = it.getDiscountMoney(discountedMoney, dateTime) }
+        return discountedMoney
+    }
+}
