+++ settings.gradle.kts
@@ -14,3 +14,5 @@ dependencyResolutionManagement {
 }
 rootProject.name = "android-movie-ticket"
 include(":app")
+include(":lib")
+include(":domain")
